% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_IMCthreshViewer.R
\encoding{UTF-8}
\name{shiny_IMCthreshViewer}
\alias{shiny_IMCthreshViewer}
\title{shiny_IMCthreshViewer}
\usage{
shiny_IMCthreshViewer()
}
\value{
app
}
\description{
App
}
\examples{
#shiny_IMCthreshViewer()
}
